+++ changelog.md
@@ -14,5 +14,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Setup Docker based build system.
 - Add default api endpoint `info`
 - Setup command based pattern using verdeter
+- Add an http error handling mecanism
 
 [unreleased]: https://github.com/ditrit/badaas/blob/main/changelog.md#unreleased